refactor(chat): quality pass — reader race hardening, markdown block extraction, perf-kind single source
Thermo dual-review synthesis (2 rounds + final recheck):
- localSyncStatusReader: forced reads bypass the shared in-flight
  promise, last-issued-wins settle ordering (seq + failuresAtIssue),
  eviction of a doomed older in-flight promise on newer settle, and a
  test-reset generation stamp; Settings mount reads are forced.
- MarkdownBody/MarkdownBlock/WorkspacePathLink/MosaicRenderContext
  moved to chatMarkdownBlock.tsx; AssistantTextBody extracted with a
  direct import (no Body prop injection); split-scan cache owned by
  useSplitRevealed in useRevealedText.
- Perf event kinds: one as-const source in perfLog with the derived
  type and isPerfEventKind validator; perfIpc consumes it.
- Reveal hook: dead cleanup effect removed, document/intersection
  visibility split into separate axes, docstring corrections.
- dev-shared: log fd 0o600, truncate failure no longer kills logging.
