+++ apps/desktop/src/renderer/lib/localSyncStatusReader.ts
@@ -1,4 +1,4 @@
-import type { SyncGetStatusArgs, SyncRoleSnapshot } from "../../shared/types";
+import type { SyncRoleSnapshot } from "../../shared/types";
 
 /**
  * Shared reader for `ade.sync.getLocalStatus`.
@@ -56,6 +56,17 @@ type ReaderState = {
   lastError: unknown;
   consecutiveFailures: number;
   nextAttemptAtMs: number;
+  /** Monotonic id stamped on every issued request. */
+  issuedSeq: number;
+  /** Highest request id that has already written the health bookkeeping. */
+  lastSettledSeq: number;
+  /** Request id currently occupying `inFlight`; 0 when none. */
+  inFlightSeq: number;
+  /**
+   * Bumped by the test reset so a read left dangling by a prior test can
+   * never settle into the fresh state. Production never bumps this.
+   */
+  generation: number;
 };
 
 const state: ReaderState = {
@@ -64,23 +75,77 @@ const state: ReaderState = {
   lastError: null,
   consecutiveFailures: 0,
   nextAttemptAtMs: 0,
+  issuedSeq: 0,
+  lastSettledSeq: 0,
+  inFlightSeq: 0,
+  generation: 0,
 };
+
+/**
+ * The identity a request carries from issue to settle.
+ *
+ * Since `force` lets a second read run alongside the shared one, two requests
+ * can be in flight at once and can finish out of order — a forced read
+ * typically beats the older shared read that is sitting at the 30s IPC timeout.
+ * Settling is therefore last-ISSUED-wins, not last-to-finish: `seq` orders the
+ * requests, and `failuresAtIssue` is the failure count this request saw when it
+ * started, so two concurrent failures record one failed generation instead of
+ * two.
+ */
+type IssuedRequest = { seq: number; failuresAtIssue: number; generation: number };
+
+function issueRequest(): IssuedRequest {
+  state.issuedSeq += 1;
+  return {
+    seq: state.issuedSeq,
+    failuresAtIssue: state.consecutiveFailures,
+    generation: state.generation,
+  };
+}
+
+/** True when a newer request has already settled, making this result stale. */
+function isStale(request: IssuedRequest): boolean {
+  return request.generation !== state.generation || request.seq < state.lastSettledSeq;
+}
+
+/**
+ * A settle from a newer request proves any older promise still occupying
+ * `inFlight` is doomed to a stale answer (typically a read stuck at the 30s
+ * IPC timeout). Evict it so later non-forced callers issue a fresh read
+ * instead of coalescing onto it; its own `.finally` no-ops once evicted.
+ */
+function evictOlderInFlight(request: IssuedRequest): void {
+  if (state.inFlight !== null && state.inFlightSeq < request.seq) {
+    state.inFlight = null;
+    state.inFlightSeq = 0;
+  }
+}
 
 /** Test seam; production always uses the real clock. */
 let now: () => number = () => Date.now();
 
-function settleHealthy(snapshot: SyncRoleSnapshot): void {
+function settleHealthy(request: IssuedRequest, snapshot: SyncRoleSnapshot): void {
+  if (isStale(request)) return;
+  state.lastSettledSeq = request.seq;
   state.lastSnapshot = snapshot;
   state.lastError = null;
   state.consecutiveFailures = 0;
   state.nextAttemptAtMs = 0;
+  evictOlderInFlight(request);
 }
 
-function settleUnhealthy(snapshot: SyncRoleSnapshot | null, error: unknown): void {
+function settleUnhealthy(
+  request: IssuedRequest,
+  snapshot: SyncRoleSnapshot | null,
+  error: unknown,
+): void {
+  if (isStale(request)) return;
+  state.lastSettledSeq = request.seq;
   state.lastSnapshot = snapshot;
   state.lastError = snapshot ? null : error;
-  state.consecutiveFailures += 1;
+  state.consecutiveFailures = request.failuresAtIssue + 1;
   state.nextAttemptAtMs = now() + localSyncStatusBackoffMs(state.consecutiveFailures);
+  evictOlderInFlight(request);
 }
 
 /**
@@ -92,35 +157,48 @@ function settleUnhealthy(snapshot: SyncRoleSnapshot | null, error: unknown): voi
  * re-thrown, which keeps the existing "Couldn't load connection details" UX.
  */
 export function readLocalSyncStatus(
-  args?: SyncGetStatusArgs,
   options?: { force?: boolean },
 ): Promise<SyncRoleSnapshot> {
-  if (state.inFlight) return state.inFlight;
-  // `force` is for reads a person asked for (a Try again click, a read right
-  // after a device mutation). Those are bounded by the user, so the backoff
-  // window must not swallow them.
+  // `force` is for reads a person asked for (a Try again click, the initial
+  // open of Connections, a read right after a device mutation). Those are
+  // bounded by the user, so neither the backoff window nor an already in-flight
+  // read — which may have been issued before the thing they just changed — may
+  // swallow them.
+  if (state.inFlight && !options?.force) return state.inFlight;
   if (!options?.force && state.consecutiveFailures > 0 && now() < state.nextAttemptAtMs) {
     return state.lastSnapshot
       ? Promise.resolve(state.lastSnapshot)
       : Promise.reject(state.lastError);
   }
+  const issued = issueRequest();
   const request = Promise.resolve()
-    .then(() => window.ade.sync.getLocalStatus(args))
+    .then(() => window.ade.sync.getLocalStatus())
     .then(
       (snapshot) => {
-        if (isDegradedLocalSyncSnapshot(snapshot)) settleUnhealthy(snapshot, null);
-        else settleHealthy(snapshot);
+        if (isDegradedLocalSyncSnapshot(snapshot)) settleUnhealthy(issued, snapshot, null);
+        else settleHealthy(issued, snapshot);
         return snapshot;
       },
       (error: unknown) => {
-        settleUnhealthy(null, error);
+        settleUnhealthy(issued, null, error);
         throw error;
       },
     )
     .finally(() => {
-      if (state.inFlight === request) state.inFlight = null;
+      if (state.inFlight === request) {
+        state.inFlight = null;
+        state.inFlightSeq = 0;
+      }
     });
-  state.inFlight = request;
+  // A forced read never displaces the shared in-flight promise: later
+  // non-forced callers keep coalescing onto the read that was already running.
+  // Its result still settles the shared health bookkeeping above, last ISSUED
+  // request winning — so a stale read finishing after a newer one cannot undo
+  // the newer answer, and a newer settle evicts the older shared promise.
+  if (state.inFlight === null) {
+    state.inFlight = request;
+    state.inFlightSeq = issued.seq;
+  }
   return request;
 }
 
@@ -131,5 +209,10 @@ export function resetLocalSyncStatusReaderForTests(clock?: () => number): void {
   state.lastError = null;
   state.consecutiveFailures = 0;
   state.nextAttemptAtMs = 0;
+  state.issuedSeq = 0;
+  state.lastSettledSeq = 0;
+  state.inFlightSeq = 0;
+  // Invalidate any read a prior test left dangling; it can never settle here.
+  state.generation += 1;
   now = clock ?? (() => Date.now());
 }
